requirements.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.10.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
environs==7.1.0
django-cors-headers==3.2.1
